directory and regex asset
